exported geocoding function to a utility script and also implemented the logic into the get university endpoint
